+++ VBPieChart/Classes/UIColor+HexColor.h
@@ -16,14 +16,8 @@
 
 
 @interface UIColor (HexColor)
-
-- (NSString *) hexString;
-
 + (UIColor *) colorWithHex:(int)color;
-+ (UIColor *) colorWithHexRed:(int)red green:(char)green blue:(char)blue alpha:(char)alpha;
-
 + (UIColor *) colorWithHexString:(NSString *)hexString;
-+ (UIColor *) colorWithIntegerRed:(float)red green:(float)green blue:(float)blue alpha:(float)alpha;
 @end