+++ VBPieChart/Classes/VBPieChart.m
@@ -23,7 +23,6 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
     CGPoint moveP;
     __strong NSMutableArray *_chartValues;
 }
-@property (nonatomic, strong) NSMutableArray *chartsData;
 @property (nonatomic) double radius;
 @property (nonatomic) double holeRadius;
 @property (nonatomic, weak) VBPiePiece *hitLayer;
@@ -32,7 +31,7 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
 @property (nonatomic) VBPieChartAnimationOptions animationOptions;
 @property (nonatomic) double animationDuration;
 
-@property (nonatomic, strong) NSMutableArray *pieceArray;
+@property (nonatomic, strong) NSMutableArray *piecesArray;
 
 @end
 
@@ -57,8 +56,12 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
 - (void) removeLable;
 
 - (void) setData:(VBPiePieceData*)data;
+- (VBPiePieceData*)data;
+
+- (CATextLayer *)label;
 
 @property (nonatomic, copy) void (^endAnimationBlock)(void);
+
 @end
 
 
@@ -85,7 +88,6 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
 }
 
 - (void)configure {
-    self.chartsData = [NSMutableArray array];
     
     self.startAngle = 0;
     self.length = M_PI*2;
@@ -133,89 +135,46 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
 }
 
 
-- (UIColor*) defaultColors:(NSInteger)index {
-    
-    // It's better then just a random
-
-    float deltaRed = 0;
-    float deltaGreen = 0;
-    float deltaBlue = 0;
-    float alpha = 1;
-    
-    if (index > 7) {
-        deltaRed = (arc4random() % 500)/2555.0f;
-        deltaGreen = (arc4random() % 500)/2555.0f;
-        deltaBlue = (arc4random() % 500)/2555.0f;
-        index = index % 8;
-    }
-    
-    switch (index) {
-        case 0:
-            return [UIColor colorWithRed:0.31+deltaRed green:0.74+deltaGreen blue:0.91+deltaBlue alpha:alpha];
-        case 1:
-            return [UIColor colorWithRed:0.33+deltaRed green:0.17+deltaGreen blue:0.48+deltaBlue alpha:alpha];
-        case 2:
-            return [UIColor colorWithRed:0.81+deltaRed green:0.61+deltaGreen blue:0.02+deltaBlue alpha:alpha];
-        case 3:
-            return [UIColor colorWithRed:0.43+deltaRed green:0.02+deltaGreen blue:0.46+deltaBlue alpha:alpha];
-        case 4:
-            return [UIColor colorWithRed:0.55+deltaRed green:0.77+deltaGreen blue:0.53+deltaBlue alpha:alpha];
-        case 5:
-            return [UIColor colorWithRed:0.00+deltaRed green:0.51+deltaGreen blue:0.08+deltaBlue alpha:alpha];
-        case 6:
-            return [UIColor colorWithRed:0.84+deltaRed green:0.66+deltaGreen blue:0.81+deltaBlue alpha:alpha];
-        case 7:
-            return [UIColor colorWithRed:0.73+deltaRed green:0.20+deltaGreen blue:0.63+deltaBlue alpha:alpha];
-            
-        default:
-            return [UIColor colorWithRed:(arc4random() % 255)/255.0f green:(arc4random() % 255)/255.0f blue:(arc4random() % 255)/255.0f alpha:alpha];
-    }
-}
-
-- (void) setValues:(NSDictionary*)values {
-    if (![values isKindOfClass:[NSDictionary class]]) {
-        [NSException raise:@"VBPieChartException" format:@"values needs to be specified with NSDictionary"];
-    }
-    for (NSNumber *key in [values allKeys]) {
-        NSInteger index = [key intValue];
-        VBPiePieceData *data = _chartsData[index];
-        data.value = values[key];
-        _chartValues[index][@"value"] = values[key];
-    }
-    [self _refreshCharts];
-}
-
 - (void) setValue:(NSNumber*)value pieceAtIndex:(NSInteger)index {
     if (![value isKindOfClass:[NSNumber class]]) {
         [NSException raise:@"VBPieChartException" format:@"value needs to be NSNumber"];
     }
-    VBPiePieceData *data = _chartsData[index];
-    data.value = value;
+    VBPiePiece *piece = _piecesArray[index];
+    piece.data.value = value;
     _chartValues[index][@"value"] = value;
     [self _refreshCharts];
 }
 
 
 - (void) removePieceAtIndex:(NSInteger)index {
-    VBPiePiece *piece = (VBPiePiece *)_pieceArray[index];
-    [piece removeLable];
     
-    VBPiePieceData *data = _chartsData[index];
-    data.value = @(0);
+    if ([CATransaction completionBlock]) {
+        void (^completionBlock2)(void) = [CATransaction completionBlock];
+        [CATransaction setCompletionBlock:nil];
+        completionBlock2();
+    }
+    
+    __block VBPiePiece *piece = _piecesArray[index];
+    [piece removeLable];
+    piece.data.value = @(0);
     [self _refreshCharts];
+    [_chartValues removeObjectAtIndex:index];
     
     void (^completionBlock)(void)  = ^(void){
-        [_chartValues removeObjectAtIndex:index];
-        [_chartsData removeObjectAtIndex:index];
-        [piece removeFromSuperlayer];
-        [_pieceArray removeObjectAtIndex:index];
         
-        for (NSInteger i = index; i < _chartsData.count; i++) {
-            VBPiePieceData *data = _chartsData[i];
-            data.index = i;
+        if (piece.animationKeys.count) {
+            [piece removeAllAnimations];
+        }
+        
+        [piece removeFromSuperlayer];
+        [_piecesArray removeObject:piece];
+        
+        for (NSInteger i = index; i < _piecesArray.count; i++) {
+            VBPiePiece *piece = _piecesArray[i];
+            piece.data.index = i;
         }
     };
-    
+
     if ([CATransaction disableActions]) {
         completionBlock();
     } else {
@@ -235,6 +194,7 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
     mutableChartValue[@"value"] = @0;
     
     [_chartValues insertObject:mutableChartValue atIndex:index];
+    [_piecesArray insertObject:[[VBPiePiece alloc] init] atIndex:index];
     _presentWithAnimation = NO;
     [self _updateCharts];
     [self setValue:value pieceAtIndex:index];
@@ -244,19 +204,17 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
 - (void) _refreshCharts {
     
     double fullValue = 0;
-    for (VBPiePieceData *data in _chartsData) {
-        fullValue += fabs([data.value doubleValue]);
+    for (VBPiePiece *piece in _piecesArray) {
+        fullValue += fabs([piece.data.value doubleValue]);
     }
     
     CGFloat onePrecent = fullValue*0.01;
     CGFloat onePrecentOfChart = _length*0.01;
     CGFloat start = _startAngle;
     
-    for (VBPiePieceData *data in _chartsData) {
-        CGFloat pieceValuePrecents = fabs([data.value doubleValue])/onePrecent;
+    for (VBPiePiece *piece in _piecesArray) {
+        CGFloat pieceValuePrecents = fabs([piece.data.value doubleValue])/onePrecent;
         CGFloat pieceChartValue = onePrecentOfChart*pieceValuePrecents;
-        
-        VBPiePiece *piece = (VBPiePiece *)_pieceArray[data.index];
         
         [piece _animateToAngle:pieceChartValue startAngle:start];
     
@@ -265,65 +223,18 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
 }
 
 
-- (double) _recreateChartsData {
-    
-    self.chartsData = [NSMutableArray array];
-    
-    double fullValue = 0;
-    
-    NSUInteger index = 0;
-    for (NSDictionary *object in _chartValues) {
-        
-        VBPiePieceData *data;
-        BOOL created = NO;
-        if ([_chartsData count] > index) {
-            data = [_chartsData objectAtIndex:index];
-        } else {
-            data = [[VBPiePieceData alloc] init];
-            created = YES;
-        }
-        data.index = index;
-        
+- (double) _allValuesSum {
+    double allValues = 0;
+    for (id object in _chartValues) {
         if ([object isKindOfClass:[NSDictionary class]]) {
             NSDictionary *dict = (NSDictionary*)object;
-            data.name = dict[@"name"];
-            data.value = dict[@"value"];
-            if (!dict[@"color"]) {
-                data.color = [self defaultColors:index];
-            } else {
-                data.color = dict[@"color"];
-            }
-            
-            if (!dict[@"labelColor"]) {
-                data.labelColor = [UIColor whiteColor];
-            } else {
-                data.labelColor = dict[@"labelColor"];
-            }
-            
-            if (dict[@"strokeColor"]) {
-                data.strokeColor = dict[@"strokeColor"];
-            }
-            
-            
-            data.accent = [dict[@"accent"] boolValue];
+            allValues += fabs([dict[@"value"] doubleValue]);
         } else {
-            data.value = (NSNumber*)object;
-            if (created) {
-                data.color = [self defaultColors:index];
-            }
+            allValues += fabs([object doubleValue]);
         }
-        
-        if (created) {
-            [_chartsData addObject:data];
-        }
-        
-        fullValue += fabs([data.value doubleValue]);
-        index++;
     }
-    
-    return fullValue;
+    return allValues;
 }
-
 
 - (void) _updateCharts {
 
@@ -331,76 +242,110 @@ static __inline__ CGFloat CGPointDistanceBetweenTwoPoints(CGPoint point1, CGPoin
         return;
     }
     
-    // Clean old layers
-    NSArray *arraySublayers = [NSArray arrayWithArray:self.layer.sublayers];
-    for (CALayer *l in arraySublayers) {
-        [l removeFromSuperlayer];
+    if (!_piecesArray) {
+        _piecesArray = [NSMutableArray array];
     }
-    [_pieceArray removeAllObjects];
+    NSMutableArray *piecesArray = [_piecesArray mutableCopy];
+    [_piecesArray removeAllObjects];
     
     // init temp variables
-    CGRect rect = CGRectMake(0, 0, self.frame.size.width, self.frame.size.height);
+    CGRect rect = self.bounds;
     
-    double fullValue = [self _recreateChartsData];
+    double allValues = [self _allValuesSum];
     
     
-    CGFloat onePrecent = fullValue*0.01;
+    CGFloat onePrecent = allValues*0.01;
     CGFloat onePrecentOfChart = _length*0.01;
     CGFloat start = _startAngle;
 
-    _pieceArray = [NSMutableArray array];
     
-    for (VBPiePieceData *data in _chartsData) {
+    NSInteger index = 0;
+    for (NSDictionary *options in _chartValues) {
+
+        VBPiePiece *piece = nil;
+        if (piecesArray.count > 0) {
+            piece = piecesArray.firstObject;
+            [piecesArray removeObjectAtIndex:0];
+        } else {
+            piece = [[VBPiePiece alloc] init];
+        }
+        
+        VBPiePieceData *data = [VBPiePieceData pieceDataWith:options];
+        data.index = index;
         
         CGFloat pieceValuePrecents = fabs([data.value doubleValue])/onePrecent;
-        CGFloat pieceChartValue = onePrecentOfChart*pieceValuePrecents;
-        
-        VBPiePiece *piece = [[VBPiePiece alloc] init];
+        CGFloat pieceAngle = onePrecentOfChart*pieceValuePrecents;
+
         [piece setFrame:rect];
-        
         [piece setLabelsPosition:_labelsPosition];
         [piece setLabelBlock:_labelBlock];
-        piece->_value = pieceValuePrecents;
         piece->_innerRadius = _radius;
         piece->_outerRadius = _holeRadius;
         
         [piece setData:data];
         
-        [piece pieceAngle:pieceChartValue start:start];
+        [piece pieceAngle:pieceAngle start:start];
         
         if (_presentWithAnimation) {
+            [CATransaction setDisableActions:YES];
             [piece setHidden:YES];
+            [CATransaction setDisableActions:NO];
         }
         [piece setAnimationDuration:_animationDuration];
         [piece setAnimationOptions:_animationOptions];
         
-        [self.layer addSublayer:piece];
-        [_pieceArray addObject:piece];
-        start += pieceChartValue;
+        if (!piece.superlayer) {
+            [self.layer addSublayer:piece];
+        }
+        if (![_piecesArray containsObject:piece]) {
+            [_piecesArray addObject:piece];
+        }
+        start += pieceAngle;
+        
+        index++;
     }
+    
+    
+    // remove old pieces
+    for (VBPiePiece *piece in piecesArray) {
+        [piece removeFromSuperlayer];
+        [piece removeLable];
+    }
+    
+    [self _runAnimation];
+}
+
+
+- (void) _runAnimation {
     
     // if was selected present with animation
     if (_presentWithAnimation) {
         
         if (_animationOptions & VBPieChartAnimationGrowthAll || _animationOptions & VBPieChartAnimationGrowthBackAll || _animationOptions & VBPieChartAnimationFan) {
             
-            for (VBPiePiece *piece in _pieceArray) {
+            for (VBPiePiece *piece in _piecesArray) {
                 [piece _animate];
             }
             
         } else {
             
-            for (NSInteger i = 0, len = _pieceArray.count; i < len; i++) {
-                VBPiePiece *piece = (VBPiePiece *)_pieceArray[i];
+            for (NSInteger i = 0, len = _piecesArray.count; i < len; i++) {
+                VBPiePiece *piece = (VBPiePiece *)_piecesArray[i];
+                [CATransaction setDisableActions:YES];
+                [piece.label setHidden:YES];
+                [CATransaction setDisableActions:NO];
                 if (i+1 < len) {
-                    __block VBPiePiece *blockPiece = (VBPiePiece *)_pieceArray[i+1];
+                    __block VBPiePiece *blockPiece = (VBPiePiece *)_piecesArray[i+1];
                     [piece setEndAnimationBlock:^{
                         [blockPiece _animate];
                     }];
                 }
             }
             
-            VBPiePiece *piece = (VBPiePiece *)_pieceArray.firstObject;
+            VBPiePiece *piece = (VBPiePiece *)_piecesArray.firstObject;
+            [CATransaction setDisableActions:YES];
+            [piece.label setHidden:YES];
+            [CATransaction setDisableActions:NO];
             [piece _animate];
         }